requirements.txt
PySide6>=6.5.0
numpy>=1.24.0
Pillow>=9.5.0
